feat: UserUsecase 구현

+++ data/repository-implementation/src/main/java/com/data/repository_implementation/UpAuthRepositoryImpl.kt
@@ -26,8 +26,8 @@ class UpAuthRepositoryImpl @Inject constructor(
 
     override suspend fun signUp(
         oauthToken: String,
-        mainRegionId: Long,
-        interestedRegionIds: List<Long>,
+        mainRegionId: Int,
+        interestedRegionIds: List<Int>,
         nickname: String,
         agreements: List<Agreement>
     ): SignUpResult {
