+++ domain/repository-interface/src/main/java/com/domain/repository_interface/UpAuthRepository.kt
@@ -10,8 +10,8 @@ interface UpAuthRepository {
     suspend fun login(oAuthToken: String): LoginResult
     suspend fun signUp(
         oauthToken: String,
-        mainRegionId: Long,
-        interestedRegionIds: List<Long>,
+        mainRegionId: Int,
+        interestedRegionIds: List<Int>,
         nickname: String,
         agreements: List<Agreement>
     ): SignUpResult
