+++ domain/usecase/src/main/java/com/domain/usecase/UpAuthUseCase.kt
@@ -12,8 +12,8 @@ interface UpAuthUseCase {
     suspend fun login(oAuthToken: String): LoginResult
     suspend fun signUp(
         oauthToken: String,
-        mainRegionId: Long,
-        interestedRegionIds: List<Long>,
+        mainRegionId: Int,
+        interestedRegionIds: List<Int>,
         nickname: String,
         agreements: List<Agreement>
     ): SignUpResult
@@ -30,8 +30,8 @@ class UpAuthUseCaseImpl @Inject constructor(
 
     override suspend fun signUp(
         oauthToken: String,
-        mainRegionId: Long,
-        interestedRegionIds: List<Long>,
+        mainRegionId: Int,
+        interestedRegionIds: List<Int>,
         nickname: String,
         agreements: List<Agreement>
     ): SignUpResult {
